✅ pass one more case

+++ 46UnitFractions/unit_fractions.c
@@ -52,7 +52,7 @@ int main()
     int ptr = 0;
     for (int i = n+1; i <= n * 2; ++i) {
       // bulk of the logic
-      float b = (float)(n*i) / (float)(i-n);
+      double b = (double)(n*i) / (double)(i-n);
       fprintf(stderr, "b is %f", b);
       if (ceilf(b) == b) {
         alist[ptr] = i;
